Move code to src dir
