+++ lib/src/animation_gesture/animated_page_dragger.dart
@@ -2,8 +2,8 @@ import 'dart:async';
 import 'dart:ui';
 
 import 'package:flutter/material.dart';
-import 'package:intro_views_flutter/Constants/constants.dart';
-import 'package:intro_views_flutter/Models/slide_update_model.dart';
+import 'package:intro_views_flutter/src/constants/constants.dart';
+import 'package:intro_views_flutter/src/models/slide_update_model.dart';
 
 /// This class provides the animation controller
 /// used when then user stops dragging and page
