+++ lib/src/animation_gesture/page_dragger.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:intro_views_flutter/Constants/constants.dart';
-import 'package:intro_views_flutter/Models/slide_update_model.dart';
+import 'package:intro_views_flutter/src/constants/constants.dart';
+import 'package:intro_views_flutter/src/models/slide_update_model.dart';
 
 /// This class is used to get user gesture and work according to it.
 class PageDragger extends StatefulWidget {
