+++ lib/src/animation_gesture/page_reveal.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:intro_views_flutter/Clipper/circular_reveal_clipper.dart';
+import 'package:intro_views_flutter/src/clipper/circular_reveal_clipper.dart';
 
 /// This class reveals the next page in the circular form.
 class PageReveal extends StatelessWidget {
