+++ lib/src/models/page_bubble_view_model.dart
