+++ lib/src/models/pager_indicator_view_model.dart
@@ -1,5 +1,5 @@
-import 'package:intro_views_flutter/Constants/constants.dart';
-import 'package:intro_views_flutter/Models/page_view_model.dart';
+import 'package:intro_views_flutter/src/constants/constants.dart';
+import 'package:intro_views_flutter/src/models/page_view_model.dart';
 
 /// View model for page indicator.
 class PagerIndicatorViewModel {
