+++ lib/src/ui/page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:intro_views_flutter/Models/page_view_model.dart';
+import 'package:intro_views_flutter/src/models/page_view_model.dart';
 
 /// This is the class which contains the Page UI.
 class Page extends StatelessWidget {
